[RF][HS3] Merge HS3 into one executable
The `testHS3SimultaneousFit` was pretty small anyway, and having only a
single executable saves linking time.
